requirements.txt
transformers
sentence-transformers
faiss-cpu
datasets
matplotlib
numpy
torch
tqdm
scikit-learn 